Ubuntu 22.04: Follow systemd to respect localhost
The default systemd shipped in Ubuntu 22.04 has the "feature" to respect
the string "localhost". That is to said, if the content of /etc/hostname
is "localhost", systemd will respect this setting and keep it as-is
a.k.a.  "localhost". This feature of systemd is introduced since systemd
v248.

By making the default an empty string, we can trigger systemd to fetch a
host names from DHCP as what the conventional way does.

Launchpad: #1991662