Fix RHEL9.6 OVS download function
RHEL9.6 requires doca-openvswitch-ipsec to be installed seperately.
The function responsible for downloading the package does download it
but creates a hang right after blocking the execution of the script.

This change changes the function to look for a single package per every
call where a pattern is provided (meaning it will download the first
package that matches the pattern).

Signed-off-by: Moshe Sheena <msheena@nvidia.com>